Rename to isMatrixRtcMode.ts
