+++ packages/react-sdk/src/lib/index.ts
@@ -21,9 +21,9 @@ export { findForegroundColor } from './findForegroundColor';
 export { acquireImageUrl, releaseImageUrl } from './imageBlobCache';
 export { isDefined } from './isDefined';
 export { isInfiniteCanvasMode } from './isInfiniteCanvasMode';
+export { isMatrixRtcMode } from './isMatrixRtcMode';
 export { setLocale } from './locale';
 export * from './matrix';
-export { isMatrixRtcMode } from './matrixRtcMode';
 export { findColor, useColorPalette } from './useColorPalette';
 export type { Color } from './useColorPalette';
 export { FontsLoadedContextProvider, useFontsLoaded } from './useFontsLoaded';
