+++ packages/react-sdk/src/lib/isMatrixRtcMode.ts
